requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dataclasses-json==0.6.7
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
isodate==0.7.2
marshmallow==3.26.1
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
PyJWT==2.9.0
python-decouple==3.8
python-youtube==0.9.7
requests==2.32.3
requests-oauthlib==1.3.0
sqlparse==0.5.3
typing-inspect==0.9.0
typing_extensions==4.13.2
urllib3==2.4.0
